Add release installation instead of just directory

+++ tests/test.bats
@@ -12,15 +12,23 @@ setup() {
 }
 
 teardown() {
-  cd ${TESTDIR}
+  cd ${TESTDIR} || ( printf "unable to cd to ${TESTDIR}\n" && exit 1 )
   ddev delete -Oy ${DDEV_SITENAME}
-  rm -rf ${TESTDIR}
+  [ "${TESTDIR}" != "" ] && rm -rf ${TESTDIR}
 }
 
-@test "basic installation" {
+@test "install from directory" {
   cd ${TESTDIR} || ( printf "unable to cd to ${TESTDIR}\n" && exit 1 )
   echo "# ddev get ${DIR} with project ${PROJNAME} in ${TESTDIR} ($(pwd))" >&3
   ddev get ${DIR}
+  ddev restart
+  ddev exec "curl -s elasticsearch:9200" | grep "${PROJNAME}-elasticsearch"
+}
+
+@test "install from release" {
+  cd ${TESTDIR} || ( printf "unable to cd to ${TESTDIR}\n" && exit 1 )
+  echo "# ddev get drud/ddev-elasticsearch with project ${PROJNAME} in ${TESTDIR} ($(pwd))" >&3
+  ddev get drud/ddev-elasticsearch
   ddev restart
   ddev exec "curl -s elasticsearch:9200" | grep "${PROJNAME}-elasticsearch"
 }
